refact(frontend): split ConfigPanel sections into components/config/
Extract the six render*Section() closures into their own files:
Teams/Overlay/Position/Buttons/Behavior/Links. Each section owns its
own useI18n() call so ConfigPanel stops having to thread labels through.
LinksSection also owns its own copiedKey state and the copyToClipboard
helper. ConfigPanel shrinks from 498 -> 217 lines and is now pure
layout + save/refresh orchestration.
