+++ frontend/src/components/config/BehaviorSection.jsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { useI18n } from '../../i18n';
+
+export default function BehaviorSection({ settings, setSetting }) {
+  const { t, lang, setLanguage, languages } = useI18n();
+  return (
+    <div className="config-section-behavior">
+      <div className="config-switch-row">
+        <label className="config-switch-label">
+          <input type="checkbox" checked={settings.autoHide}
+            onChange={(e) => setSetting('autoHide', e.target.checked)} />
+          {t('behavior.autoHide')}
+        </label>
+      </div>
+      {settings.autoHide && (
+        <div className="config-range-row">
+          <label className="config-label">{t('behavior.hideAfter', { value: settings.autoHideSeconds })}</label>
+          <input type="range" min={1} max={15} step={1} value={settings.autoHideSeconds}
+            onChange={(e) => setSetting('autoHideSeconds', Number(e.target.value))} className="config-range" />
+        </div>
+      )}
+      <div className="config-switch-row">
+        <label className="config-switch-label">
+          <input type="checkbox" checked={settings.autoSimple}
+            onChange={(e) => setSetting('autoSimple', e.target.checked)} />
+          {t('behavior.autoSimple')}
+        </label>
+      </div>
+      {settings.autoSimple && (
+        <div className="config-switch-row" style={{ paddingLeft: '1.5rem' }}>
+          <label className="config-switch-label">
+            <input type="checkbox" checked={settings.autoSimpleOnTimeout}
+              onChange={(e) => setSetting('autoSimpleOnTimeout', e.target.checked)} />
+            {t('behavior.fullOnTimeout')}
+          </label>
+        </div>
+      )}
+
+      <div className="config-separator" />
+      <div className="config-field-row">
+        <label className="config-label">{t('lang.label')}</label>
+        <select className="config-select" value={lang} onChange={(e) => setLanguage(e.target.value)}>
+          {languages.map((l) => (
+            <option key={l} value={l}>{l === 'en' ? 'English' : l === 'es' ? 'Español' : l}</option>
+          ))}
+        </select>
+      </div>
+    </div>
+  );
+}
